PlayerMovement and Rotation
Added a new class in common called PlayerMovement since player movement is not using acceleration and so on. (We can use this class for the enemys aswell if they don't need to use acceleration). WASD controls the movement and the arrow keys changes the rotation (note if the arrow keys are not pressed the movement keys are also controlling the rotation).